exercises/lab_02_mcp_advanced/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_agent"]

[project]
name = "mcp-agent"
version = "0.1.0"
description = "Advanced MCP Agent Labs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "markdownify>=1.2.2",
    "mcp>=1.25.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]

[project.scripts]
coding-agent = "mcp_agent.coding_agent.main:main"
deep-agent = "mcp_agent.deep_agent.main:main"

[dependency-groups]
dev = [
    "ipython>=9.9.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.uv]
package = true

[tool.pyrefly]

